backend/ml_service/requirements.txt
# ============================================================
# ml_service/requirements.txt
# Install with: pip install -r requirements.txt
# ============================================================

# Web framework
fastapi==0.110.0
uvicorn==0.27.0
python-multipart==0.0.9

# Image processing (preprocessing pipeline)
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
scikit-image==0.22.0

# Printed OCR (Tesseract wrapper)
# ALSO install Tesseract binary:
#   Windows: https://github.com/UB-Mannheim/tesseract/wiki
#   Linux:   sudo apt-get install tesseract-ocr tesseract-ocr-hin tesseract-ocr-mar
pytesseract==0.3.10

# Handwritten Text Recognition (TrOCR — Microsoft, runs locally)
transformers==4.38.2
torch==2.2.1
torchvision==0.17.1

# Document classification (YOLOv8 + EfficientNet)
ultralytics==8.1.24

# Data augmentation (for model training)
albumentations==1.3.1

# Text comparison engine (replaces Gemini arbitrator)
rapidfuzz==3.6.1
spacy==3.7.4
# After install, download spaCy model:
#   python -m spacy download en_core_web_sm
